+++ src/interfaces/wasi/tests/mod.rs
@@ -1,8 +1,23 @@
+// Tests performances:
+// Install:
+// curl https://wasmtime.dev/install.sh -sSf | bash
+// cargo install cargo-wasi
+// Launch bench
+// cargo wasi test --release --features wit -- --nocapture single_test
+
+use std::time::Instant;
+
 use super::wit_generation::{
     decrypt, delegate_user_decryption_key, encrypt, generate_master_key,
     generate_user_decryption_key, rotate_attributes, Attribute, Policy, PolicyAxis,
 };
-use crate::core::policy::ap;
+use crate::{
+    core::policy::ap,
+    interfaces::wasi::wit_generation::{
+        create_encryption_cache, destroy_encryption_cache, encrypt_hybrid_block,
+        encrypt_hybrid_header,
+    },
+};
 
 #[test]
 fn access_policy_test() {
@@ -19,7 +34,7 @@ fn single_test() {
         primary_axis: super::wit_generation::PolicyAxis {
             name: "Departments".to_string(),
             attributes: vec![
-                "RnD".to_string(),
+                "R&D".to_string(),
                 "HR".to_string(),
                 "MKG".to_string(),
                 "FIN".to_string(),
@@ -27,13 +42,13 @@ fn single_test() {
             hierarchical: false,
         },
         secondary_axis: PolicyAxis {
-            name: "Security_Level".to_string(),
+            name: "Security Level".to_string(),
             attributes: vec![
-                "level_1".to_string(),
-                "level_2".to_string(),
-                "level_3".to_string(),
-                "level_4".to_string(),
-                "level_5".to_string(),
+                "level 1".to_string(),
+                "level 2".to_string(),
+                "level 3".to_string(),
+                "level 4".to_string(),
+                "level 5".to_string(),
             ],
             hierarchical: true,
         },
@@ -47,7 +62,7 @@ fn single_test() {
 
     let level_4_mkg_fin_delegate = generate_user_decryption_key(
         mk.private_key,
-        Some("Departments::MKG & Security_Level::level_4".to_string()),
+        Some("Departments::MKG && Security Level::level 4".to_string()),
         mk.policy_serialized.clone(),
     )
     .unwrap();
@@ -56,7 +71,7 @@ fn single_test() {
         mk.delegation_key.clone(),
         level_4_mkg_fin_delegate,
         mk.policy_serialized.clone(),
-        Some("Departments::MKG & Security_Level::level_3".to_string()),
+        Some("Departments::MKG && Security Level::level 3".to_string()),
     );
 
     let updated_policy = rotate_attributes(
@@ -68,24 +83,55 @@ fn single_test() {
     )
     .unwrap();
 
-    let ciphertext = encrypt(
-        "plaintext".to_string(),
-        mk.public_key,
-        vec![
-            Attribute {
-                axis_name: "Departments".to_string(),
-                attribute: "MKG".to_string(),
-            },
-            Attribute {
-                axis_name: "Security_Level".to_string(),
-                attribute: "level_3".to_string(),
-            },
-        ],
-        updated_policy,
-        // mk.policy_serialized,
-    )
-    .unwrap();
+    let uid = vec![0_u8; 32];
+
+    let loops = 10;
+    let before = Instant::now();
+    let abe_attributes = vec![
+        Attribute {
+            axis_name: "Departments".to_string(),
+            attribute: "MKG".to_string(),
+        },
+        Attribute {
+            axis_name: "Security Level".to_string(),
+            attribute: "level 3".to_string(),
+        },
+    ];
+
+    let mut ciphertext = Vec::<u8>::new();
+    for _i in 0..loops {
+        ciphertext = encrypt(
+            "plaintext".to_string(),
+            mk.public_key.clone(),
+            abe_attributes.clone(),
+            updated_policy.clone(),
+            // mk.policy_serialized,
+            uid.clone(),
+        )
+        .unwrap();
+    }
+    let avg_time = before.elapsed().as_micros() / loops;
+    println!("\navg time (no cache)\t: {} micro seconds", avg_time);
 
     let cleartext = decrypt(super_delegate, ciphertext).unwrap();
     assert_eq!("plaintext".to_string(), cleartext);
+
+    let cache_handle = create_encryption_cache(mk.public_key, updated_policy).unwrap();
+    let before = Instant::now();
+    for _i in 0..loops {
+        let header =
+            encrypt_hybrid_header(abe_attributes.clone(), cache_handle, uid.clone()).unwrap();
+        encrypt_hybrid_block(
+            "plaintext".to_string(),
+            header.symmetric_key,
+            uid.clone(),
+            0,
+        )
+        .unwrap();
+    }
+
+    let avg_time = before.elapsed().as_micros() / loops;
+    println!("avg time (with cache)\t: {} micro seconds", avg_time);
+
+    destroy_encryption_cache(cache_handle).unwrap();
 }